+++ src/main/java/org/cicirello/search/evo/AdaptiveGeneration.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2023 Vincent A. Cicirello
+ * Copyright (C) 2002-2024 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -22,6 +22,7 @@ package org.cicirello.search.evo;
 
 import org.cicirello.math.rand.EnhancedSplittableGenerator;
 import org.cicirello.search.internal.RandomnessFactory;
+import org.cicirello.search.internal.ReferenceValidator;
 import org.cicirello.search.operators.CrossoverOperator;
 import org.cicirello.search.operators.MutationOperator;
 import org.cicirello.util.Copyable;
@@ -49,12 +50,8 @@ final class AdaptiveGeneration<T extends Copyable<T>> implements Generation<T> {
   private final EnhancedSplittableGenerator generator;
 
   AdaptiveGeneration(MutationOperator<T> mutation, CrossoverOperator<T> crossover) {
-    if (mutation == null) {
-      throw new NullPointerException("mutation must be non-null");
-    }
-    if (crossover == null) {
-      throw new NullPointerException("crossover must be non-null");
-    }
+    ReferenceValidator.nullCheck(mutation);
+    ReferenceValidator.nullCheck(crossover);
     this.mutation = mutation;
     this.crossover = crossover;
     generator = RandomnessFactory.createEnhancedSplittableGenerator();