+++ src/main/java/org/cicirello/search/evo/BasePopulation.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2023 Vincent A. Cicirello
+ * Copyright (C) 2002-2024 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -22,6 +22,7 @@ package org.cicirello.search.evo;
 
 import java.util.ArrayList;
 import org.cicirello.search.ProgressTracker;
+import org.cicirello.search.internal.ReferenceValidator;
 import org.cicirello.search.operators.Initializer;
 import org.cicirello.util.Copyable;
 
@@ -79,9 +80,10 @@ abstract class BasePopulation {
       if (n < 1) {
         throw new IllegalArgumentException("population size n must be positive");
       }
-      if (initializer == null || f == null || selection == null || tracker == null) {
-        throw new NullPointerException("passed a null object for a required parameter");
-      }
+      ReferenceValidator.nullCheck(initializer);
+      ReferenceValidator.nullCheck(f);
+      ReferenceValidator.nullCheck(selection);
+      ReferenceValidator.nullCheck(tracker);
       this.initializer = initializer;
       this.selection = selection;
 
@@ -251,9 +253,10 @@ abstract class BasePopulation {
       if (n < 1) {
         throw new IllegalArgumentException("population size n must be positive");
       }
-      if (initializer == null || f == null || selection == null || tracker == null) {
-        throw new NullPointerException("passed a null object for a required parameter");
-      }
+      ReferenceValidator.nullCheck(initializer);
+      ReferenceValidator.nullCheck(f);
+      ReferenceValidator.nullCheck(selection);
+      ReferenceValidator.nullCheck(tracker);
       this.initializer = initializer;
       this.selection = selection;