+++ src/main/java/org/cicirello/search/evo/OnlyAlwaysMutateGeneration.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2023 Vincent A. Cicirello
+ * Copyright (C) 2002-2024 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -20,6 +20,7 @@
 
 package org.cicirello.search.evo;
 
+import org.cicirello.search.internal.ReferenceValidator;
 import org.cicirello.search.operators.MutationOperator;
 import org.cicirello.util.Copyable;
 
@@ -36,9 +37,7 @@ final class OnlyAlwaysMutateGeneration<T extends Copyable<T>> implements Generat
   private final MutationOperator<T> mutation;
 
   OnlyAlwaysMutateGeneration(MutationOperator<T> mutation) {
-    if (mutation == null) {
-      throw new NullPointerException("mutation must be non-null");
-    }
+    ReferenceValidator.nullCheck(mutation);
     this.mutation = mutation;
   }