+++ src/main/java/org/cicirello/search/hc/SteepestDescentHillClimber.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2022 Vincent A. Cicirello
+ * Copyright (C) 2002-2024 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -22,6 +22,7 @@ package org.cicirello.search.hc;
 
 import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.SolutionCostPair;
+import org.cicirello.search.internal.ReferenceValidator;
 import org.cicirello.search.operators.Initializer;
 import org.cicirello.search.operators.IterableMutationOperator;
 import org.cicirello.search.operators.MutationIterator;
@@ -71,9 +72,8 @@ public final class SteepestDescentHillClimber<T extends Copyable<T>>
       Initializer<T> initializer,
       ProgressTracker<T> tracker) {
     super(initializer, tracker);
-    if (problem == null || mutation == null) {
-      throw new NullPointerException();
-    }
+    ReferenceValidator.nullCheck(problem);
+    ReferenceValidator.nullCheck(mutation);
     this.mutation = mutation;
     pOpt = problem;
     pOptInt = null;
@@ -96,9 +96,8 @@ public final class SteepestDescentHillClimber<T extends Copyable<T>>
       Initializer<T> initializer,
       ProgressTracker<T> tracker) {
     super(initializer, tracker);
-    if (problem == null || mutation == null) {
-      throw new NullPointerException();
-    }
+    ReferenceValidator.nullCheck(problem);
+    ReferenceValidator.nullCheck(mutation);
     this.mutation = mutation;
     pOptInt = problem;
     pOpt = null;
@@ -119,9 +118,8 @@ public final class SteepestDescentHillClimber<T extends Copyable<T>>
       IterableMutationOperator<T> mutation,
       Initializer<T> initializer) {
     super(initializer, new ProgressTracker<T>());
-    if (problem == null || mutation == null) {
-      throw new NullPointerException();
-    }
+    ReferenceValidator.nullCheck(problem);
+    ReferenceValidator.nullCheck(mutation);
     this.mutation = mutation;
     pOpt = problem;
     pOptInt = null;
@@ -142,9 +140,8 @@ public final class SteepestDescentHillClimber<T extends Copyable<T>>
       IterableMutationOperator<T> mutation,
       Initializer<T> initializer) {
     super(initializer, new ProgressTracker<T>());
-    if (problem == null || mutation == null) {
-      throw new NullPointerException();
-    }
+    ReferenceValidator.nullCheck(problem);
+    ReferenceValidator.nullCheck(mutation);
     this.mutation = mutation;
     pOptInt = problem;
     pOpt = null;