+++ src/main/java/org/cicirello/search/ss/IterativeSampling.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2023 Vincent A. Cicirello
+ * Copyright (C) 2002-2024 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -22,6 +22,7 @@ package org.cicirello.search.ss;
 
 import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.SolutionCostPair;
+import org.cicirello.search.internal.ReferenceValidator;
 import org.cicirello.search.operators.Initializer;
 import org.cicirello.search.problems.IntegerCostOptimizationProblem;
 import org.cicirello.search.problems.OptimizationProblem;
@@ -59,9 +60,7 @@ public final class IterativeSampling<T extends Copyable<T>> extends AbstractStoc
   public IterativeSampling(
       OptimizationProblem<T> problem, Initializer<T> initializer, ProgressTracker<T> tracker) {
     super(problem, tracker);
-    if (initializer == null) {
-      throw new NullPointerException();
-    }
+    ReferenceValidator.nullCheck(initializer);
     this.initializer = initializer;
   }
 
@@ -79,9 +78,7 @@ public final class IterativeSampling<T extends Copyable<T>> extends AbstractStoc
       Initializer<T> initializer,
       ProgressTracker<T> tracker) {
     super(problem, tracker);
-    if (initializer == null) {
-      throw new NullPointerException();
-    }
+    ReferenceValidator.nullCheck(initializer);
     this.initializer = initializer;
   }